requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.1
django-cors-middleware==1.3.1
django-extensions==2.1.0
djangorestframework==3.9.2
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
gunicorn==19.9.0
isort==4.3.20
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.0
pylint==2.3.1
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1
requests==2.22.0
ruamel.yaml==0.15.96
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
six==1.12.0
typed-ast==1.3.5
whitenoise==4.1.2
wrapt==1.11.1
django-cors-headers==3.0.1
django-filter==2.1.0
python-twitter==3.5
facebook-sdk==3.1.0
google-auth==1.6.3    

